Introduce MaybeStringTypeTrait
